refactor(ai-agent): move remote_agent files to manager/remote/
