+++ crates/aionui-ai-agent/src/lib.rs
@@ -16,11 +16,9 @@ pub mod cli_process;
 pub mod factory;
 pub mod first_message_injector;
 pub mod idle_scanner;
+pub mod manager;
 pub mod nanobot_agent;
 pub mod openclaw;
-pub mod remote_agent;
-pub mod remote_agent_routes;
-pub mod remote_agent_service;
 pub mod skill_manager;
 pub mod stream_event;
 pub mod task_manager;
@@ -40,11 +38,11 @@ pub use backend_protocol_sink::BackendProtocolSink;
 pub use cli_process::CliAgentProcess;
 pub use factory::{AgentFactoryDeps, build_agent_factory};
 pub use idle_scanner::start_idle_scanner;
+pub use manager::remote::{
+    RemoteAgentConfig, RemoteAgentManager, RemoteAgentRouterState, RemoteAgentService, remote_agent_routes,
+};
 pub use nanobot_agent::NanobotAgentManager;
 pub use openclaw::OpenClawAgentManager;
-pub use remote_agent::{RemoteAgentConfig, RemoteAgentManager};
-pub use remote_agent_routes::{RemoteAgentRouterState, remote_agent_routes};
-pub use remote_agent_service::RemoteAgentService;
 pub use skill_manager::{
     AcpSkillManager, SkillDefinition, SkillIndex, build_skills_index_text, build_system_instructions,
     build_system_instructions_with_skills_index, detect_skill_load_request, prepare_first_message,
